Reorganise Analytics code ready for new API
